+++ src/main.cpp
@@ -26,7 +26,7 @@ using namespace blackjack;
 
 namespace {
 
-constexpr const char* kVersion = "1.2.0";
+constexpr const char* kVersion = "1.2.1";
 
 // ----- tiny argument helpers ------------------------------------------------